Duplicate the error_generator in message expectations
Historically we've always passed this in containing the target
object in the proxy which helps us with generating the correct
error messages, however we then mutate its opts which leaks upwards
to the proxy. In addition if we set a custom error message this
is then shared by all message expectations which is certainly
not desired.

Fixes rspec/rspec#234